+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- AppBar color
+
 ## [0.1.2] - 2022-08-02
 ### Changed
 - Tasks with a status "app not found" are removed after 1h as they provide no value in history